requirements.txt
slack_bolt==1.18.0
slack_sdk==3.21.3
dlt
dlt[bigquery]
google-api-python-client
google-cloud-storage
websockets
google-cloud-bigquery-storage
pandas==2.2.3
db-dtypes==1.3.0
flask